refactor: move new organisms to dedicated folder
