--- src/components/organisms/NewPaneManager/index.ts
@@ -1,1 +1,0 @@
-export {default} from './NewPaneManager';
