+++ src/components/organisms/index.ts
@@ -8,6 +8,7 @@ export {default as CreateProjectModal} from './CreateProjectModal';
 export {default as FileTreePane} from './FileTreePane';
 export {default as FiltersPresetModal} from './FiltersPresetModal';
 export {default as GitCloneModal} from './GitCloneModal';
+export {default as GitOpsView} from './GitOpsView';
 export {default as GitPane} from './GitPane';
 export {default as HelmPane} from './HelmPane';
 export {default as ImagesPane} from './ImagesPane';
@@ -18,7 +19,6 @@ export {default as LazyDrawer} from './LazyDrawer';
 export {default as LocalResourceDiffModal} from './LocalResourceDiffModal';
 export {default as MessageBox} from './MessageBox';
 export {default as NavigatorPane} from './NavigatorPane';
-export {default as NewPaneManager} from './NewPaneManager';
 export {default as NewResourceWizard} from './NewResourceWizard';
 export {default as NotificationsManager} from './NotificationsManager';
 export {default as PageHeader} from './PageHeader';
@@ -28,9 +28,11 @@ export {default as QuickSearchActions} from './QuickSearchActions';
 export {default as RenameEntityModal} from './RenameEntityModal';
 export {default as RenameResourceModal} from './RenameResourceModal';
 export {default as ReplaceImageModal} from './ReplaceImageModal';
+export {default as SearchPane} from './SearchPane';
 export {default as SettingsManager} from './SettingsManager';
 export {default as TemplateManagerPane} from './TemplateManagerPane';
 export {default as TemplateModal} from './TemplateModal';
 export {default as TerminalPane} from './TerminalPane';
 export {default as UpdateNotice} from './UpdateNotice';
 export {default as ValidationPane} from './ValidationPane';
+export {RecentProjectsPage, StartProjectPage} from './StartProjectPane';
