+++ src/components/organisms_new/NavigatorPane/NavigatorPane.tsx
@@ -1,0 +1,5 @@
+const NavigatorPane = () => {
+  return null;
+};
+
+export default NavigatorPane;
