+++ src/components/organisms_new/NavigatorPane/index.ts
@@ -1,0 +1,1 @@
+export {default} from './NavigatorPane';
