+++ src/components/organisms_new/PaneManager/PaneManager.styled.tsx
