+++ src/components/organisms_new/PaneManager/index.ts
@@ -1,0 +1,1 @@
+export {default} from './PaneManager';
