+++ src/components/organisms_new/index.ts
@@ -1,0 +1,2 @@
+export {default as PaneManager} from './PaneManager';
+export {default as NavigatorPane} from './NavigatorPane';
